+++ framework/Furion.Extras.DatabaseAccessor.Dapper/Locators/IDapperLocator.cs
@@ -1,0 +1,9 @@
+﻿namespace Dapper
+{
+    /// <summary>
+    /// Dapper ORM 定位器
+    /// </summary>
+    public interface IDapperLocator
+    {
+    }
+}