+++ framework/Furion/JsonSerialization/Providers/SystemTextJsonSerializerProvider.cs
@@ -1,4 +1,5 @@
 ﻿using Furion.DependencyInjection;
+using System.Text.Encodings.Web;
 using System.Text.Json;
 
 namespace Furion.JsonSerialization
@@ -17,7 +18,8 @@ namespace Furion.JsonSerialization
         /// <returns></returns>
         public string Serialize(object value, object jsonSerializerOptions = null)
         {
-            return JsonSerializer.Serialize(value, jsonSerializerOptions as JsonSerializerOptions);
+            return JsonSerializer.Serialize(value, (jsonSerializerOptions as JsonSerializerOptions)
+                ?? GetDefaultJsonSerializerOptions());
         }
 
         /// <summary>
@@ -29,7 +31,22 @@ namespace Furion.JsonSerialization
         /// <returns></returns>
         public T Deserialize<T>(string json, object jsonSerializerOptions = null)
         {
-            return JsonSerializer.Deserialize<T>(json, jsonSerializerOptions as JsonSerializerOptions);
+            return JsonSerializer.Deserialize<T>(json, (jsonSerializerOptions as JsonSerializerOptions)
+                ?? GetDefaultJsonSerializerOptions());
+        }
+
+        /// <summary>
+        /// 默认配置
+        /// </summary>
+        /// <returns></returns>
+        private static JsonSerializerOptions GetDefaultJsonSerializerOptions()
+        {
+            return new JsonSerializerOptions
+            {
+                WriteIndented = true,   // 缩进
+                Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,  // 中文乱码
+                PropertyNameCaseInsensitive = true  // 忽略大小写
+            };
         }
     }
 }